requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt